+++ app/routes/musings.$id/route.tsx
@@ -1,0 +1,4 @@
+export default function MusingIdRoute() {
+    return <div>MusingIdRoute</div>;
+    
+}
