+++ react-router.config.ts
@@ -3,5 +3,8 @@ import type { Config } from "@react-router/dev/config";
 export default {
   // Config options...
   // Server-side render by default, to enable SPA mode set this to `false`
+
+  // https://reactrouter.com/start/framework/data-loading#static-data-loading
+
   ssr: true,
 } satisfies Config;
